==> recomendador-de-anime/01_Descarga_Dataset.ipynb <==
## Descargar Dataset 

https://www.kaggle.com/hernan4444/anime-recommendation-database-2020

Para una guía de como hacer una descarga de datasets de kaggle:

https://github.com/Kaggle/kaggle-api#download-dataset-files

https://www.ankushchoubey.com/download_kaggle/

#!pip3 install kaggle

!kaggle

data_dir = "anime-data"
dataset_name =  "hernan4444/anime-recommendation-database-2020"
!mkdir $data_dir
!kaggle datasets download -p {data_dir}/ --unzip {dataset_name}
