==> recomendador-de-anime/05_Probando_Recomendaciones.ipynb <==
%store -r
%store

# Interactuando con campañas

import time
from time import sleep
import json
import numpy as np
import pandas as pd
pd.set_option('display.max_columns', 500)     # Make sure we can see all of the columns
pd.set_option('display.max_rows', 50)         # Keep the output on one page
pd.set_option('display.max_colwidth', 200)  


from datetime import datetime
import uuid
import random

import boto3
import botocore
from botocore.exceptions import ClientError
import pandas as pd

personalize = boto3.client('personalize')
personalize_runtime = boto3.client('personalize-runtime')

# Establish a connection to Personalize's event streaming
personalize_events = boto3.client(service_name='personalize-events')

anime = pd.read_csv(data_dir + '/anime.csv',index_col='MAL_ID')
anime.sample(1)

# Funcion para verificar la existencia del anime

def get_anime_by_id(anime_id, df=anime):
    try:
        return df.loc[int(anime_id)]['Name']
    except:
        return "Error obtaining title"

def get_animerow_by_id(anime_id, df=anime):
    try:
        return df.loc[int(anime_id)]
    except:
        return "Error obtaining title"

def get_anime_like(name, df=anime):
    try:
        found = df[df.Name.apply(lambda x: name.lower() in x.lower())].sample().index.values[0]
        return found
    except:
        return "Error no hay"

def get_anime_equal(name, df=anime):
    try:
        found = df[df.Name.apply(lambda x: name.lower() == x.lower())].sample().index.values[0]
        return found
    except:
        return "Error no hay"

# Similar Items (SIMS)

# Update DF rendering
pd.set_option('display.max_rows', 30)

def get_new_recommendations_df(recommendations_df, anime_ID):
    # Get the movie name
    anime_name = get_anime_by_id(anime_ID)
    # Get the recommendations
    get_recommendations_response = personalize_runtime.get_recommendations(
        campaignArn = sims_campaign_arn,
        itemId = str(anime_ID),
    )
    # Build a new dataframe of recommendations
    item_list = get_recommendations_response['itemList']
    recommendation_list = []
    for item in item_list:
        tanime = get_anime_by_id(item['itemId'])
        recommendation_list.append(tanime)
    new_rec_DF = pd.DataFrame(recommendation_list, columns = [anime_name])
    # Add this dataframe to the old one
    recommendations_df = pd.concat([recommendations_df, new_rec_DF], axis=1)
    return recommendations_df

sims_recommendations_df = pd.DataFrame()
animes = [get_anime_equal('fullmetal alchemist'), get_anime_equal('death note'), get_anime_equal('Hellsing Ultimate'), get_anime_equal('One Punch Man')]

for a in animes:
    sims_recommendations_df = get_new_recommendations_df(sims_recommendations_df, a)

sims_recommendations_df[:15]

# User Personalization


# Update DF rendering
pd.set_option('display.max_rows', 30)

def get_new_recommendations_df_users(recommendations_df, user_id):
    # Get the movie name
    #movie_name = get_movie_by_id(artist_ID)
    # Get the recommendations
    get_recommendations_response = personalize_runtime.get_recommendations(
        campaignArn = userpersonalization_campaign_arn,
        userId = str(user_id),
    )
    # Build a new dataframe of recommendations
    item_list = get_recommendations_response['itemList']
    recommendation_list = []
    for item in item_list:
        anime = get_anime_by_id(item['itemId'])
        recommendation_list.append(anime)
    #print(recommendation_list)
    new_rec_DF = pd.DataFrame(recommendation_list, columns = [user_id])
    # Add this dataframe to the old one
    recommendations_df = pd.concat([recommendations_df, new_rec_DF], axis=1)
    return recommendations_df

recommendations_df_users = pd.DataFrame()
#users = users_df.sample(3).index.tolist()

for user in [1,293,29738, 3990, 283878]:
    recommendations_df_users = get_new_recommendations_df_users(recommendations_df_users, user)

recommendations_df_users[:10]

# Usando Filtros

def get_new_recommendations_df_by_filter(recommendations_df, user_id, filter_arn):
    # Get the movie name
    #movie_name = get_movie_by_id(artist_ID)
    # Get the recommendations
    get_recommendations_response = personalize_runtime.get_recommendations(
        campaignArn = userpersonalization_campaign_arn,
        userId = str(user_id),
        filterArn = filter_arn
    )
    # Build a new dataframe of recommendations
    item_list = get_recommendations_response['itemList']
    recommendation_list = []
    for item in item_list:
        anime = get_anime_by_id(item['itemId'])
        recommendation_list.append(anime)
    #print(recommendation_list)
    filter_name = filter_arn.split('/')[1]
    new_rec_DF = pd.DataFrame(recommendation_list, columns = [filter_name])
    # Add this dataframe to the old one
    recommendations_df = pd.concat([recommendations_df, new_rec_DF], axis=1)
    return recommendations_df

## Veamos las recomendaciones por todas los géneros que cuentan con filtros..

usuario  = 30000

recommendations_df_shelves = pd.DataFrame()
for filter_arn in meta_filter_arns:
    recommendations_df_shelves = get_new_recommendations_df_by_filter(recommendations_df_shelves, usuario, filter_arn)

recommendations_df_shelves[:10]

El siguiente tema son los eventos en tiempo real. Personalizar tiene la capacidad de escuchar eventos de su aplicación para actualizar las recomendaciones que se muestran al usuario. Esto es especialmente útil en cargas de trabajo de medios, como videos, donde la intención de un cliente puede diferir según si lo está viendo con sus hijos o por su cuenta.

Además, los eventos que se registran a través de este sistema se almacenan hasta que se emite una llamada de eliminar, y se utilizan como datos históricos junto con los demás datos de interacción que proporcionó cuando entrena sus próximos modelos.

# Eventos en tiempo real (Event tracker)

empezamos creando un evento

response = personalize.create_event_tracker(
    name='AnimeTracker',
    datasetGroupArn=dataset_group_arn
)
print(response['eventTrackerArn'])
print(response['trackingId'])
TRACKING_ID = response['trackingId']
event_tracker_arn = response['eventTrackerArn']

session_dict = {}

def rate_anime(USER_ID, ITEM_ID,EVENT_VALUE, EVENT_TYPE='RATING' ):
    """
    Simula una calificación
    para enviarla a Amazon personalize como un evento. 
    """
    # Configure Session
    try:
        session_ID = session_dict[str(USER_ID)]
    except:
        session_dict[str(USER_ID)] = str(uuid.uuid1())
        session_ID = session_dict[str(USER_ID)]
        
    # Configure Properties:
    event = {
    "itemId": str(ITEM_ID),
    }
    event_json = json.dumps(event)
        
    # Hace la llamada
    
    personalize_events.put_events(
    trackingId = TRACKING_ID,
    userId= str(USER_ID),
    sessionId = session_ID,
    eventList = [{
        'sentAt': int(time.time()),
        'eventType': str(EVENT_TYPE),
        'eventValue': int(EVENT_VALUE),
        'properties': event_json
        }]
    )

def get_new_recommendations_df_users_real_time(user_id):
    # Get the artist name (header of column)
    # Get the recommendations (note you should have a base recommendation DF created before)
    get_recommendations_response = personalize_runtime.get_recommendations(
        campaignArn = userpersonalization_campaign_arn,
        userId = str(user_id),
    )
    # Build a new dataframe of recommendations
    item_list = get_recommendations_response['itemList']
    recommendation_list = []
    for item in item_list:
        anime = get_anime_by_id(item['itemId'])
        recommendation_list.append(anime)
    new_rec_DF = pd.DataFrame(recommendation_list, columns = ['Recomendaciones'])
    # Add this dataframe to the old one
    #recommendations_df = recommendations_df.join(new_rec_DF)
    return new_rec_DF

rating_complete = pd.read_csv(data_dir + '/rating_complete.csv', usecols=['user_id'])

rating_complete['user_id'].unique().max()

# Cualquier usuario sobre 353404 es nuevo ;)

#Cliente abre su cuenta de video streaming.
user_id = 2000000
get_new_recommendations_df_users_real_time(user_id)[:10]

# Cliente ve y califica Hunter X Hunter
anime_name = 'Fullmetal Alchemist'
nota = 9
anime_id = get_anime_equal(anime_name)
print (f'Calificando {anime_name}({anime_id}) con nota {nota}')
rate_anime(user_id, anime_id, nota)

get_new_recommendations_df_users_real_time(user_id)[:10]

# Cliente ve y califica Steins;Gate
anime_name = 'Elfen Lied'
nota = 7
anime_id = get_anime_equal(anime_name)
print (f'Calificando {anime_name}({anime_id}) con nota {nota}')
rate_anime(user_id, anime_id, nota)

get_new_recommendations_df_users_real_time(user_id)[:10]

# Cliente ve y califica ranma
anime_name = 'Naruto'
nota = 6
anime_id = get_anime_like(anime_name)
print (f'Calificando {anime_name}({anime_id}) con nota {nota}')
rate_anime(user_id, anime_id, nota)

get_new_recommendations_df_users_real_time(user_id)[:10]

# Cliente ve y califica shingeki no kyojin
anime_name = 'Hunter x Hunter (2011)'
nota = 8
anime_id = get_anime_like(anime_name)
print (f'Calificando {anime_name}({anime_id}) con nota {nota}')
rate_anime(user_id, anime_id, nota)

get_new_recommendations_df_users_real_time(user_id)[:10]
