==> recomendador-de-anime/07_Metric_attribution.ipynb <==
# Metric Attribution 
Más información: https://docs.aws.amazon.com/personalize/latest/dg/measuring-recommendation-impact.html

import boto3
import json
import pandas as pd
import datetime

# Configure the SDK to Personalize:
personalize = boto3.client('personalize')
personalize_runtime = boto3.client('personalize-runtime')
personalize_events = boto3.client('personalize-events')

## Para encontrar el dataset group arn: 
df = pd.DataFrame(personalize.list_dataset_groups()['datasetGroups'])
df

dataset_group_arn = dataset_group_arn
path_to_bucket = path_to_bucket
metric_attribution_name = metric_attribution_name
role_arn = role_arn

metrics_list = [
	{
		"eventType": "CLICK",
		"metricName": "COUNT_CLICK",
		"expression": "SAMPLECOUNT()"
	},
	{
		"eventType": "RATING",
		"metricName": "COUNT_RATING",
		"expression": "SAMPLECOUNT()"
	},
	{
		"eventType": "WATCH",
		"metricName": "COUNT_WATCH",
		"expression": "SAMPLECOUNT()"
	}
]

output_config = {
    "roleArn": role_arn,
    "s3DataDestination": {
    "path": path_to_bucket    
    }
}

response = personalize.create_metric_attribution(
name = metric_attribution_name,
datasetGroupArn = dataset_group_arn,
metricsOutputConfig = output_config,
metrics = metrics_list
)
metric_attribution_arn = response['metricAttributionArn']

# Enviar eventos manualmente

session_dict = {}
import uuid

def rate_anime(USER_ID, ITEM_ID,EVENT_VALUE, EVENT_TYPE ):
    """
    Simula una calificación
    para enviarla a Amazon personalize como un evento. 
    """
    # Configure Session
    try:
        session_ID = session_dict[str(USER_ID)]
    except:
        session_dict[str(USER_ID)] = str(uuid.uuid1())
        session_ID = session_dict[str(USER_ID)]
        
    # Configure Properties:
    event = {
    "itemId": str(ITEM_ID),
    }
    event_json = json.dumps(event)
        
    # Hace la llamada
    
    personalize_events.put_events(
    trackingId = TRACKING_ID,
    userId= str(USER_ID),
    sessionId = session_ID,
    eventList = [{
        'sentAt': int(time.time()),
        'eventType': str(EVENT_TYPE),
        'eventValue': int(EVENT_VALUE),
        'properties': event_json,
        'metricAttribution': {"eventAttributionSource": "escenario_a"},
        }]
    )

TRACKING_ID = TRACKING_ID
USER_ID = USER_ID
ITEM_ID = ITEM_ID
EVENT_VALUE = INT 
EVENT_TYPE = 'RATING'|"CLICK"|"WATCH"
rate_anime(USER_ID, ITEM_ID,EVENT_VALUE, EVENT_TYPE)
